refactor: Implementa validaciones, excepciones 404 y protección transaccional

+++ src/main/java/com/yamilrafart/apirest_productos/service/ProductoServiceImpl.java
@@ -1,8 +1,10 @@
 package com.yamilrafart.apirest_productos.service;
 
 import com.yamilrafart.apirest_productos.entity.Producto;
+import com.yamilrafart.apirest_productos.exception.ResourceNotFoundException;
 import com.yamilrafart.apirest_productos.repository.ProductoRepository;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 
@@ -15,37 +17,39 @@ public class ProductoServiceImpl implements IProducto {
         this.productoRepository = productoRepository;
     }
 
-
     @Override
+    @Transactional
     public Producto save(Producto producto) {
         return productoRepository.save(producto);
     }
 
-
     @Override
+    @Transactional(readOnly = true) // readOnly = true optimiza las consultas de solo lectura
     public List<Producto> findAll() {
         return productoRepository.findAll();
     }
 
-
     @Override
+    @Transactional(readOnly = true) // readOnly = true optimiza las consultas de solo lectura
     public Producto findById(Long id) {
         return productoRepository.findById(id)
-                .orElseThrow(() -> new RuntimeException("Producto no encontrado"));
+                .orElseThrow(() -> new ResourceNotFoundException("Producto no encontrado con el ID: " + id));
     }
 
-
     @Override
+    @Transactional
     public void deleteById(Long id) {
+        if (!productoRepository.existsById(id)) {
+            throw new ResourceNotFoundException("No se puede eliminar. Producto no encontrado con el ID: " + id);
+        }
         productoRepository.deleteById(id);
     }
 
-
     @Override
+    @Transactional
     public Producto update(Producto producto) {
-
-        Producto productoBD = productoRepository.findById(producto.getId())
-                .orElseThrow(() -> new RuntimeException("Producto no encontrado"));
+        //findById para buscar y lanzar la excepción si no existe
+        Producto productoBD = this.findById(producto.getId());
 
         productoBD.setNombre(producto.getNombre());
         productoBD.setDescripcion(producto.getDescripcion());
